SB-833: `addArtifactToIndex` method added into `LayoutProvider`
