+++ strongbox-storage/strongbox-storage-core/src/main/java/org/carlspring/strongbox/io/ArtifactOutputStream.java
@@ -19,7 +19,7 @@ import org.slf4j.LoggerFactory;
  *
  * @author Sergey Bespalov
  */
-public class ArtifactOutputStream
+public abstract class ArtifactOutputStream
         extends MultipleDigestOutputStream
 {
