+++ strongbox-storage/strongbox-storage-layout-providers/strongbox-storage-maven-layout-provider/src/main/java/org/carlspring/strongbox/providers/layout/Maven2LayoutProvider.java
@@ -1,5 +1,20 @@
 package org.carlspring.strongbox.providers.layout;
 
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.attribute.BasicFileAttributes;
+import java.security.NoSuchAlgorithmException;
+import java.util.Collections;
+import java.util.stream.Stream;
+
+import javax.annotation.PostConstruct;
+import javax.inject.Inject;
+
+import org.apache.maven.artifact.Artifact;
+import org.apache.maven.artifact.repository.metadata.Metadata;
+import org.apache.maven.index.ArtifactInfo;
 import org.carlspring.maven.commons.util.ArtifactUtils;
 import org.carlspring.strongbox.artifact.coordinates.MavenArtifactCoordinates;
 import org.carlspring.strongbox.providers.io.RepositoryFileAttributes;
@@ -24,20 +39,7 @@ import org.carlspring.strongbox.storage.repository.Repository;
 import org.carlspring.strongbox.storage.search.SearchRequest;
 import org.carlspring.strongbox.storage.search.SearchResult;
 import org.carlspring.strongbox.storage.search.SearchResults;
-
-import javax.annotation.PostConstruct;
-import javax.inject.Inject;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.attribute.BasicFileAttributes;
-import java.security.NoSuchAlgorithmException;
-import java.util.Collections;
-import java.util.stream.Stream;
-
-import org.apache.maven.artifact.Artifact;
-import org.apache.maven.artifact.repository.metadata.Metadata;
-import org.apache.maven.index.ArtifactInfo;
+import org.carlspring.strongbox.util.IndexContextHelper;
 import org.codehaus.plexus.util.xml.pull.XmlPullParserException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -415,6 +417,35 @@ public class Maven2LayoutProvider extends AbstractLayoutProvider<MavenArtifactCo
         artifactIndexesService.rebuildIndex(storageId, repositoryId, null);
     }
 
+    
+    @Override
+    protected void addArtifactToIndex(RepositoryPath repositoryPath)
+        throws IOException
+    {
+        Repository repository = repositoryPath.getFileSystem().getRepository();
+        Storage storage = repository.getStorage();
+        
+
+        String contextId = IndexContextHelper.getContextId(storage.getId(), repository.getId(), IndexTypeEnum.LOCAL.getType());
+        RepositoryIndexer indexer = repositoryIndexManager.getRepositoryIndexer(contextId);
+
+        Boolean artifactAttribute = (Boolean) Files.getAttribute(repositoryPath, RepositoryFileAttributes.ARTIFACT);
+        
+        if (!repository.isIndexingEnabled() || !Boolean.TRUE.equals(artifactAttribute) || indexer == null)
+        {
+            return;
+        }
+
+        String repositoryRelativePath = repositoryPath.getRepositoryRelative().toString();
+        Artifact artifact = ArtifactUtils.convertPathToArtifact(repositoryRelativePath);
+
+        File storageBasedir = new File(storage.getBasedir());
+        File artifactFile = new File(new File(storageBasedir, repository.getId()), repositoryRelativePath).getCanonicalFile();
+
+        indexer.addArtifactToIndex(repository.getId(), artifactFile, artifact);
+    }
+
+    
     @Override
     public MavenRepositoryFeatures getRepositoryFeatures()
     {
